requirements.txt
certifi
cffi==1.15.1
charset-normalizer==3.4.1
click
colorama 
cryptography==44.0.2
cycler==0.11.0
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
flit_core
fonttools==4.38.0
greenlet==3.1.1
gunicorn @ file:gunicorn-19.9.0-py2.py3-none-any.whl
idna==3.10
importlib-metadata
itsdangerous
Jinja2==3.1.6
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.5.3
numpy==1.21.6
opencv-python==4.11.0.86
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
pycparser==2.21
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.31.0
seaborn==0.12.2
six==1.17.0
SQLAlchemy==2.0.40
torch @ file:torch-1.8.0%2Bcu111-cp37-cp37m-win_amd64.whl
torchvision @ file:torchvision-0.9.0-cp37-cp37m-win_amd64.whl
tqdm==4.67.1
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wincertstore==0.2
xlrd==2.0.1
zipp 